refactor(doctor): split run_doctor into SRP submodules
The 160-line run_doctor function mixed six independent diagnostic
phases (auth, state dir, tool versions, registry connectivity, git
context, encryption) plus shared finding rendering. Extract each
phase into its own module under crates/shipper-cli/src/doctor/
so adding a new diagnostic is an additive change rather than an
edit of one long function, and so each check can be reasoned about
on its own.

Behavior preserved — output text and finding records are identical;
existing tests (doctor unit, bdd_doctor, e2e_doctor) pass unchanged.
